internal/factory/container: make SpecAddNamespaces platform-specific
This allows us to set the network namespace of a FreeBSD container.
